Refactor tow model for easy playing around
Parameters and match information are R data structures that are passed to WebPPL.
Lazy is a memoized function of both a person and a match, so we can later ask if a person was lazy on a particular match.
Sample return statements show the kinds of queries you can include.
ggplot at the end to visualize the posterior.